refactor: reorganize test structure by moving tests from lib.rs to a dedicated tests.rs file, improving code organization and maintainability; add new test snapshots for various equity scenarios
